requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2
pydantic
pydantic[email]
python-dotenv
gradio
groq